Revert "y-script: Bump script variables limit to bigger value"
This reverts commit f21112b7e4a35a81c687e7c3a9eb14657497de1f.

+++ src/Engine/Script.h
@@ -1810,7 +1810,7 @@ protected:
 /**
  * Strong typed collection of values for script.
  */
-template<typename T, typename I = Uint16>
+template<typename T, typename I = Uint8>
 class ScriptValues : ScriptValuesBase
 {
 public:
